LayoutTests/http/tests/websocket/tests/hybi/contentextensions/display-none-worker-expected.txt

CONSOLE MESSAGE: worker onopen
CONSOLE MESSAGE: worker onmessage reply sent
CONSOLE MESSAGE: PASS worker onclose
This should not be hidden.
